embedding/requirements.txt
streamlit
python-dotenv
PyPDF2
faiss-cpu
langchain
langchain-community
langchain-openai
langchain-huggingface
langchain-text-splitters
sentence-transformers
